refactor: RequestStatus 폴더 이동 (common)
