+++ src/store-requests/DTO/update-store-request.dto.ts
@@ -1,5 +1,5 @@
 import { IsEnum, IsNumber, IsString } from "class-validator"
-import { RequestStatus } from "../../stores/entities/request-status.enum"
+import { RequestStatus } from "../../common/request-status.enum"
 
 export class UpdateStoreRequestDTO {
     @IsEnum(RequestStatus)
